+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.44.0] - 2021-06-06
+
+### Added
+
+- Lightning value 4 value: Added option to generate invoice to recharge the wallet used for sending value.  
+  For LND users, the _macaroon_ used for authentication must include the permission `invoices:write`
+
 ## [1.43.0] - 2021-05-28
 
 ### Added
@@ -47,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Player option to play next or previous episodes in reverse order (i.e. play older episodes next instead of newer)
 
-[Unreleased]: https://github.com/podStation/podStation/compare/v1.40.1...HEAD
+[Unreleased]: https://github.com/podStation/podStation/compare/v1.44.0...HEAD
+[1.44.0]: https://github.com/podStation/podStation/compare/v1.43.0...v1.44.0
 [1.43.0]: https://github.com/podStation/podStation/compare/v1.42.0...v1.43.0
 [1.42.0]: https://github.com/podStation/podStation/compare/v1.41.0...v1.42.0
 [1.41.0]: https://github.com/podStation/podStation/compare/v1.40.1...v1.41.0